11/27/2018
Added edit method in data class.
Added privilege field to provider class
Added a totalFees field to provider class. I think the fees should be type double or float in the Service class,but I made totalFees of type int.  I will change it to double when the fees in the service class is updated to type double or float.