11/29/2018 Commit
Provider unit test changed. I won't be testing build report method, it's too complicated to automate. The testing of the build report method will be considered an integration test/ system test.

Data class and provider class are complete.  They are the final version, and the provider class writes to a file. I've tested it in main, and if you want to see what the report looks like, look at the text file "temp.txt" and "tempt2.txt" or call the displayAll() method to see it on screen.

"temp.txt": shows a provider report with services already in a service list
"temp2.txt":  shows a provider report with no services in the service list

My intellij was writing the file to my test and providerTest folder when the file name is "test.txt" without a file path.  See if it does the same for you.

I left total fees data member an int value, Java's set precision to two digits after the decimal of a float requires another class object that I don't want to mess with.  And this will cause major changes in the Service class.

If I'm having a hard time uploading the text files on github, I'll post my temp files on slack.